Created abstract class for context menu view model classes to inherit from
